+++ app/routes/reminders/new.js
@@ -5,20 +5,4 @@ export default Ember.Route.extend({
   model() {
     return Ember.Object.create();
   },
-
-  actions: {
-    createReminder(reminderForm) {
-      let stringDate = reminderForm.get('date');
-      if(stringDate) {
-        reminderForm.set('date', new Date(stringDate));
-      } else {
-        reminderForm.set('date', new Date());
-      }
-      let reminderProps = reminderForm.getProperties('title', 'date', 'notes');
-      let record = this.get('store').createRecord('reminder', reminderProps);
-      record.save().then((reminder) => {
-          this.transitionTo('reminders.reminder', reminder);
-      });
-    }
-  }
 });
