+++ app/templates/components/reminder-form.hbs
@@ -1,0 +1,15 @@
+<form>
+  <label>
+    Title
+    {{input class='spec-title-input' value=reminder.title}}
+  </label>
+  <label>
+    Date
+    {{input type='date' class='spec-date-input' value=reminder.date}}
+  </label>
+  <label>
+    Notes
+    {{textarea class='spec-note-input' value=reminder.notes}}
+  </label>
+  <button type='submit' class='spec-submit-reminder'>Submit</button>
+</form>
